rental/serializers/CarSerializer.py
from rest_framework import serializers
from rental.models import Car


class CarSerializer(serializers.ModelSerializer):
    class Meta:
        model = Car
        fields = ['id', 'make', 'model', 'is_rented']
